fix: update signup-as-student and fix prisma client

+++ apps/web/src/components/project/edit/edit-project-form.tsx
@@ -62,8 +62,10 @@ export function EditProjectForm({ projectId }: { projectId: string }) {
         variant: "contained",
         color: "error",
       },
-    }).then(() => {
-      handleDelete();
+    }).then((value) => {
+      if (value.confirmed) {
+        handleDelete();
+      }
     });
   };
